+++ src/pkg.ts
@@ -1,6 +1,6 @@
 /** @format */
 
 export default {
-  version: '0.2.4-beta.6',
-  name: 'aliyun-pds-js-sdk',
+  "version": "0.2.4-beta.7",
+  "name": "aliyun-pds-js-sdk"
 }